Tests for "/date" added
