+++ src/main/kotlin/controller/HelloController.kt
@@ -40,26 +40,4 @@ class HelloController {
         model["message"] = message
         return "welcome"
     }
-	
-	/**
-     * This function act as the handler of the HelloController.
-     *
-     * **Note** 
-     * 
-     * The annotation `@GetMapping` allows us to handle all the GET petitions to the path `/date`
-	 * using this controller.
-     *
-     * @param model collection with the data used to update the view (template)
-     * @return the template with the updated information
-     */
-    @GetMapping("/date")
-    fun date(model: MutableMap<String, Any>): String {
-		//Get the current date with format "yyyy-MM-dd"
-		val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
-		val currentDate = LocalDateTime.now().format(formatter)
-		// This is used to associate the variable "currentDate" of the template date 
-		// with the current date.
-		model["currentDate"] = currentDate
-        return "date"
-    }
 }
